==> Tumor_evolution_Chinese_CCRCC/Rcode/1_CHKIRC_Subtype.Rmd <==
---
title: "1_CHKIRC_Subtype_Analysis"
author: "`r config::get(file = 'config.yml')$document$author`"
date: "`r Sys.Date()`"
output:
  pdf_document:
    keep_tex: true
    latex_engine: pdflatex
    fig_caption: yes
    highlight: haddock
    number_sections: yes

knit: (
  function(inputFile, encoding) {
    config=config::get(file = 'config.yml') 
    
    pSubTitle <- "1_CHKIRC_Subtype_Analysis.pdf"
    
    if (!dir.exists(config$Report)) {
      dir.create(config$Report, recursive = TRUE)
    }
    
    base_output_dir <- config$Report
    
    DIR=file.path(base_output_dir)
    
    rmarkdown::render(
      input = inputFile,
      encoding = encoding,
      output_file = pSubTitle,
      output_dir = DIR)
    }
  )
---





```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F,cache = F,warning=F,message = F,cache=F, include = F)
```


```{r,include=F}
rm(list = ls(all.names = TRUE))

```

```{r}
# packages installation
#install.packages("factoextra")
#install.packages("umap")
#install.packages("pheatmap")
#install.packages("cgwtools")
```


```{r, include=FALSE}
# Define your own config
getwd()
config <- config::get(file = 'config.yml')
# 
Working_path=config$Path
email=config$Email
Authtoken=config$Token_file
knitr::opts_chunk$set(echo = F,warnings =F, message=F,include=F)
options(knitr.table.format = 'markdown',encoding = 'UTF-8',warnings =F, message=F ) 
# Check and create output directory if it doesn't exist
output_dir <- file.path(Working_path, "Data_preparation")
if (!dir.exists(output_dir)) {
  dir.create(output_dir, recursive = TRUE)
}
Reference_path=config$Reference_path

dir(Reference_path)
# load function
source("Rfunction.R" ) 
```


```{r,include=F}
Rcode_path=getwd()
Path=Working_path
Rdata_path=file.path(Path,"Rdata")
Rawdata_path=file.path(Path,"Rawdata")
Out_table_path=file.path(Path,"Out","table")
dir.create(Out_table_path,recursive = T)
#NC_Rawdata_path=file.path(Path,"Raw_data","NC_230")
```


```{r,include=F}
library(dplyr)

library(ggplot2)
load(file.path(Reference_path,"Human_GSEA","GO_KEGGP_WP_C2CGP_Reactome_pathway_list.rdata"))
cgwtools::lsdata(file.path(Reference_path,"Human_GSEA","GO_KEGGP_WP_C2CGP_Reactome_pathway_list.rdata"))
GMT_files=list.files(file.path(Reference_path,"Human_GSEA"),pattern="v7.5.1.symbols.gmt.txt",full.names = T)%>%.[grepl("c2|h",.)]
```



```{r}
cbp1 <- c("#999999", "#E69F00", "#56B4E9", "#009E73",
          "#F0E442", "#0072B2", "#D55E00", "#CC79A7")

# The palette with black:
cbp2 <- c("#000000", "#E69F00", "#56B4E9", "#009E73",
          "#F0E442", "#0072B2", "#D55E00", "#CC79A7")






```



#  Tumor/normal ratio in CHKIRC data (54 samples)


```{r}
#Proteomics data tumor vs normal log2 ratio :  /Rawdata/Proteomics/54_Proteomics_T&N_ratio.txt", Proteomics data singscore: Rawdata_path/Proteomics/54_Proteomics_singscore.txt
  
dir(Rawdata_path)
dir(file.path(Rawdata_path,"CHKIRC"))
Pro_log_data=read.table(file.path(Rawdata_path,"CHKIRC","Proteomics","54_Proteomics_T&N_ratio.txt"),fill=T)
Pro_log_data_rm_na=Pro_log_data%>%na.omit()%>%as.data.frame()
clinical_54_data_2=read.table(file.path(Rawdata_path, "CHKIRC","Clinic","54_Clinical_data.txt"),header = T,check.names = F)
Invasion_anno=clinical_54_data_2%>%dplyr::filter(Capsular_invasion=="Y" )
```

```{r}
if( !file.exists(file.path(Rawdata_path,"CHKIRC","Proteomics","54_Singscore.rdata"))){
# Download GSEA signature data from https://www.gsea-msigdb.org/gsea/msigdb
 
  all_pathways2=c()
  for (x in 1:length(GMT_files)){
      data=fgsea::gmtPathways(GMT_files[x])
      all_pathways2=c(all_pathways2,data)
    }

  set.seed(12345)
  KIRC_singscore_data=Singscore_get(Pro_log_data_rm_na,all_pathways2)
  KIRC_singscore_data=as.data.frame( KIRC_singscore_data)
  rownames(KIRC_singscore_data)=colnames(Pro_log_data_rm_na)
  KIRC_singscore_data=t(KIRC_singscore_data)%>%as.data.frame()
  save( KIRC_singscore_data,file=file.path(Rawdata_path,"CHKIRC","Proteomics","54_Singscore.rdata"  ))
 
 } else {
   
   load (file.path(Rawdata_path,"CHKIRC","Proteomics","54_Singscore.rdata"  ))
 }
KIRC_singscore_data_2=t(KIRC_singscore_data)%>%as.data.frame()

```


## Using kmeans to cluster Proteomics singscore data (top 2000 pathways with highest variance,Cluster1,CHKIRCtype2; Cluster2 and 3,CHKIRC_type1**

```{r,include=F,eval=T}
# select invasion samples for subtypes analysis
Pro_singscore=KIRC_singscore_data
Pro_singscore_Var=apply(Pro_singscore,1,var) %>%as.data.frame() %>%setNames("Var")
Pro_singscore_Var_2000=Pro_singscore %>%mutate(Var=Pro_singscore_Var$Var)%>% as.data.frame()%>%arrange(desc(Var))%>%.[1:2000,]
#Data_select=Pro_singscore %>% .[,colnames(.) %in% rownames(Pro_singscore_Var_2000)]

df=Pro_singscore_Var_2000[,1:54]%>%t()%>%as.data.frame()
```

```{r,include=T,eval=T}
set.seed(1234)
km.res=kmeans(df, 3, nstart = 25)
km.res_anno=data.frame(km.res$cluster)%>%setNames("Cluster") %>%tibble::rownames_to_column(var="Sample_ID")
```

```{r,include=T,eval=T,fig.cap="kmeans for the invasion samples"}
cluster_OUT=factoextra::fviz_cluster(km.res, df, geom = c("point"),show.clust.cent = F,ellipse.type = "convex")
 
```

```{r}
Data=df
set.seed(12345)
sig_pathways_umap=UMAP_output(Data)%>% left_join(.,km.res_anno[,c("Sample_ID","Cluster")]) %>%mutate(Cluster=as.character(Cluster))%>%mutate(Subtype=ifelse(Cluster==3,"CHKIRC_subtype2","CHKIRC_subtype1"))



subtype_color=cbp1[c(1,2,3)]
names(subtype_color)=c("1","2","3")
P1=UMAP_plot(Data_umap=sig_pathways_umap,shape_label="",color_label="Cluster",color_defination=subtype_color)

subtype_color=cbp1[c(3,5)]
names(subtype_color)=c("CHKIRC_subtype1","CHKIRC_subtype2")
P2=UMAP_plot(Data_umap=sig_pathways_umap,shape_label="",color_label="Subtype",color_defination=subtype_color)

```

**supplementary fig**

```{r,fig.width=20,fig.height=8,include=T}
cowplot::plot_grid(cluster_OUT, P1,nrow=1)
```

```{r}
Subtype1_ID=sig_pathways_umap%>%dplyr::filter(Subtype=="CHKIRC_subtype1")%>%.$Sample_ID
Subtype2_ID=sig_pathways_umap%>%dplyr::filter(Subtype=="CHKIRC_subtype2")%>%.$Sample_ID

Diff_pathways=Wilcox_test_singscore(KIRC_singscore_data,Subtype1_ID,Subtype2_ID)
Diff_pathways_p005=Diff_pathways %>%dplyr::filter(pvalue<0.05)
```

```{r}
clinical_subtype=clinical_54_data_2%>%left_join(.,sig_pathways_umap[,c("Sample_ID","Subtype")])
write.csv(clinical_subtype,file.path(Out_table_path,"CHKIRC_subtype_clinial.csv"))
```



```{r}
cat("Up pathways in Type2 vs Type1,TGF-beta: EMT,FOCAL_ADHESION,MYOGENESIS,COMPLEMENT,PI3KAKT,MTOR, save in Upregulated pathways in CHKIRC_subtype2 vs 1.csv\n" )
cat( "DN pathways: GLYCOLYSIS_AND_GLUCONEOGENESIS,FATTY_ACID_BETAOXIDATION,LONG_CHAIN_FATTY_ACID_BETAOXIDATION,HALLMARK_DNA_REPAIR in DNregulated pathways in CHKIRC_subtype2 vs 1.csv")

Diff_pathways_Hallmark_up=Diff_pathways_p005%>%dplyr::filter(log2FoldChange>0) %>%dplyr::filter(grepl("HALLMARK|WP_",SYMBOL))%>%arrange(pvalue)
write.csv(Diff_pathways_Hallmark_up,file.path(Out_table_path,"Table1,Upregulated pathways in CHKIRC_subtype2 vs 1.csv"))
```



```{r}
Diff_pathways_Hallmark_DN=Diff_pathways_p005%>%dplyr::filter(log2FoldChange<0) %>%dplyr::filter(grepl("HALLMARK|WP_",SYMBOL))%>%arrange(pvalue)
write.csv(Diff_pathways_Hallmark_DN,file.path(Out_table_path,"Table2,DNregulated pathways in CHKIRC_subtype2 vs 1.csv"))

```

## Choose most significant pathways in these pathways based on pvalue 

**"HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION" ,"WP_FOCAL_ADHESION_PI3KAKTMTORSIGNALING_PATHWAY","WP_TGFBETA_RECEPTOR_SIGNALING_IN_SKELETAL_DYSPLASIAS","HALLMARK_MYOGENESIS","WP_PI3KAKT_SIGNALING_PATHWAY" ,"WP_COMPLEMENT_SYSTEM" ,"WP_GLYCOLYSIS_AND_GLUCONEOGENESIS"   ,"WP_MITOCHONDRIAL_LONG_CHAIN_FATTY_ACID_BETAOXIDATION"**
```{r}
Compare_list=list(c("CHKIRC_subtype1" ,"CHKIRC_subtype2" ))

Subtype_pathways_data=Diff_pathways_p005%>%dplyr::filter(grepl("GLYCOLYSIS|OXIDATION|EPITHELIAL|MTOR|DNA_REPAIR|TGF|MYOGENESIS|COMPLEMENT|PI3K|AKT|PI3KAKT|APOPTOSIS",SYMBOL))%>%dplyr::filter(grepl("HALLMARK|WP_",SYMBOL))%>%arrange(pvalue)

subtype_color=cbp1[c(3,5)]
names(subtype_color)=c("CHKIRC_subtype1","CHKIRC_subtype2")
```


```{r}
Subtype_pathways=c("HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION" ,"WP_FOCAL_ADHESION_PI3KAKTMTORSIGNALING_PATHWAY","WP_TGFBETA_RECEPTOR_SIGNALING_IN_SKELETAL_DYSPLASIAS","HALLMARK_MYOGENESIS","WP_PI3KAKT_SIGNALING_PATHWAY" ,"WP_COMPLEMENT_SYSTEM" ,"WP_GLYCOLYSIS_AND_GLUCONEOGENESIS"   ,"WP_MITOCHONDRIAL_LONG_CHAIN_FATTY_ACID_BETAOXIDATION"    )

```


```{r,fig.height=25,fig.width=40,eval=T,include=F}
Norm_singscore_anno=KIRC_singscore_data [Subtype_pathways,]%>%t()%>%as.data.frame() %>%dplyr::mutate(Sample_ID=rownames(.))%>%left_join(.,sig_pathways_umap[,c("Sample_ID","Subtype")] )%>%arrange(Subtype)
Norm_singscore_anno$Subtype=factor(Norm_singscore_anno$Subtype,levels=c("CHKIRC_subtype1","CHKIRC_subtype2"))
P3=Batch_singscore_boxplot (Subtype_pathways,Norm_singscore_anno,"Subtype" ,fillby="Subtype",nCol=4,Compare_list,subtype_color )
```

 
```{r,fig.height=8,fig.width=12,include=F}
sig_pathway_anno=Norm_singscore_anno[,c("Sample_ID","Subtype")]%>%matrix.please()%>%as.data.frame()%>%setNames("Subtype")
sig_pathway_anno$Subtype=factor(sig_pathway_anno$Subtype)
Singscore_data_t_select=Norm_singscore_anno[,Subtype_pathways]%>%t()%>%as.data.frame()
colnames(Singscore_data_t_select)=Norm_singscore_anno$Sample_ID
ann_colors=list(Subtype=subtype_color)
PHEATMAP=pheatmap::pheatmap(Singscore_data_t_select,cellwigermline=5, cellheight = 15,cellwidth = 10, scale = "row",
                                    treeheight_row = 5,
                                    show_rownames = T,show_colnames = F,
                                    annotation_col= sig_pathway_anno, 
                                    annotation_colors = ann_colors,
                                    # annotation_row=Annotation,
                                    #annotation_legend=Label_def,
                                    cluster_rows = T,  cluster_cols = F,clustering_distance_rows = "euclidean")
```
 

```{r}
Data=KIRC_singscore_data [Subtype_pathways,]%>%t()%>%as.data.frame()
set.seed(12345)
sig_pathways_umap=UMAP_output(Data)%>% left_join(.,Norm_singscore_anno[,c("Sample_ID","Subtype")]) 

P4=UMAP_plot(Data_umap=sig_pathways_umap,shape_label="",color_label="Subtype",color_defination=subtype_color)

```

\newpage

```{r,fig.width=20,fig.height=6,include=T}
#cowplot::plot_grid(PHEATMAP$gtable,P4,nrow=1,labels = c("A","B"),scale=c(1,0.8))

cowplot::plot_grid(PHEATMAP$gtable,labels = c("A"))

```

```{r,fig.width=14,fig.height=6,out.width="50%",include=T}
cowplot::plot_grid(P4,grid::nullGrob() ,labels = c("B",""))

```


```{r,fig.height=18,fig.width=30,eval=T,include=T}
cowplot::plot_grid(P3,nrow=1,labels = c("C"))

```


```{r}
pheatmap_pathway_3<- function (expression_data, Gene_select,pathway_name,Annotation) {
 # Gene_select<-Pathway_list[,pathway_name]%>%unique()
  Gene_select_anno= expression_data[Gene_select,] 
  pheatmap::pheatmap(Gene_select_anno, cellwigermline= 2, cellheight =10, scale = "row",
                     treeheight_row = 5,main=pathway_name,
                     show_rownames = T,show_colnames = F,
                     annotation_col= Annotation,
                     # annotation_row=Annotation,
                     #annotation_legend=Label_def,
                     cluster_rows = T,  cluster_cols = F,clustering_distance_rows = "euclidean")
  
}

```







################################################################################

## DEPs between subtype1 and 2 (tumor/normal ratio)

```{r}
# saved in "CHKIRC_DEPs_in_subtypes.csv 
Ttest_Data=Pro_log_data_rm_na
outlist=list()
Type1_ID=sig_pathways_umap[,c("Sample_ID","Subtype")]%>%dplyr::filter(Subtype=="CHKIRC_subtype1")%>%.$Sample_ID
Type2_ID=sig_pathways_umap[,c("Sample_ID","Subtype")]%>%dplyr::filter(Subtype=="CHKIRC_subtype2")%>%.$Sample_ID
# 
# for(i in rownames(Ttest_Data)) {
#   data=Ttest_Data[i,Type1_ID]%>%as.numeric()
#   data2=Ttest_Data[i,Type2_ID]%>%as.numeric()
#   TTEST=wilcox.test(data,data2)
#   outlist[[i]]=data.frame(gene_symbol=i,pvalue=TTEST$p.value,log_a=mean(data),log_b=mean(data2))%>%mutate(log2FoldChange=log_b-log_a)
# }
# 
#  
# Ttest_out=do.call("rbind",outlist)
# 
# Ttest_out_sig=Ttest_out%>%dplyr::filter(pvalue<0.05 )%>%dplyr::filter(log2FoldChange>0.67|log2FoldChange<as.numeric(-0.67))
```


```{r}
wilcox_list=list()
for(i in rownames(Ttest_Data)) {
  data=Ttest_Data[i,Type1_ID]%>%as.numeric()
  data2=Ttest_Data[i,Type2_ID]%>%as.numeric()
  TTEST=wilcox.test(data,data2)
  outlist[[i]]=data.frame(gene_symbol=i,pvalue=TTEST$p.value,log_a=mean(data),log_b=mean(data2))%>%mutate(log2FoldChange=log_b-log_a)
}

 
wilcox_list_out=do.call("rbind",outlist)
wilcox_test_out_sig=wilcox_list_out%>%dplyr::filter(pvalue<0.05 )%>%dplyr::filter(log2FoldChange>0.26|log2FoldChange<as.numeric(-0.26))
write.csv(wilcox_test_out_sig, file.path(Rawdata_path,"CHKIRC","Proteomics","CHKIRC_DEPs_in_subtypes.csv" ))
```

### Volcano plot for all DEPs (1153,FC>1.2,up 474,DN 679,wilcox test,p<0.05)

```{r}
Valcano_plot_all_proteins=function(Data,Title){
   
  
  library(ggrepel)
  # add a column of NAs
  de=Data %>%mutate(diffexpressed=ifelse(pvalue>0.05|abs(log2FoldChange)<0.26,"NO",ifelse(log2FoldChange>0,"UP","DOWN")))

  de$delabel <- NA
 
  de$delabel[de$diffexpressed != "NO"] <- de$gene_symbol[de$diffexpressed != "NO"]
   # plot adding up all layers we have seen so far
 
  ggplot(data=de, aes(x=log2FoldChange, y=-log10(pvalue), col=diffexpressed, label=delabel)) +
    geom_point() + ggtitle(Title)+
    theme_minimal() +xlim(c(-2,2))+
    geom_text_repel() +
    scale_color_manual(values=c("DOWN"="#18499E", "NO"= "grey","UP"= "#E7211A")) +
    geom_vline(xintercept=c(-0.26, 0.26), col="red") +
    geom_hline(yintercept=-log10(0.05), col="red")
  
  
  
  
}
```


```{r,fig.width=6,fig.height=4,include=T}
Valcano_plot_all_proteins(wilcox_list_out,"All DEPs in KIRC_type2 vs KIRC_type1")
```

### supplementary data
### Volcano plot of differential expressed proteins for each pathway

```{r}
# plot volcano plot for each pathwau in significant pathways
Pathway_names=Subtype_pathways
plist=list()
for (i in Pathway_names){
 Pathway_genes=GSEA_pathway_list[[i]]
 Sig_proteins=intersect(Pathway_genes,wilcox_list_out$gene_symbol) 
 if(length(Sig_proteins)!="0") {
 Exp_data=wilcox_list_out%>%dplyr::filter(gene_symbol%in%Sig_proteins)
 plist[[i]]=Valcano_plot_NES(Exp_data,i)
 }
}
```

```{r,fig.height=24,fig.width=20,include=T,out.width="80%"}

# add title to cowplot
cowplot::plot_grid(plotlist = plist,nrow=4)+
  ggtitle("Volcano plot for each pathway in significant pathways") +
  theme(plot.title = element_text(hjust = 0.5, size = 20, face = "bold"))


```

# NES analysis subtype2 VS TYPE1(tumor/noral ratio)

```{r}
# NES data file "CHKIRC_NES_subtype2_Vs_type1.csv"

wilcox_list_out_p005=wilcox_list_out%>%dplyr::filter(pvalue<0.05)
wiocox_list=list (subtype_diff=wilcox_list_out_p005)
Out=perform_fgsea_analysis(wiocox_list,GMT_files)
FGSEA_subtype_list=Out[[1]]%>%as.data.frame()%>%dplyr::filter(pathway %in%Subtype_pathways)%>%arrange(NES)%>%mutate(Database=ifelse(grepl("HALLMARK",pathway),"Hallmark_dataset","C2_dataset"))%>%mutate(State=ifelse(pathway%in%c("WP_GLYCOLYSIS_AND_GLUCONEOGENESIS"        , "WP_MITOCHONDRIAL_LONG_CHAIN_FATTY_ACID_BETAOXIDATION"),"low pathways","high pathways"))

FGSEA_subtype_list$pathway=factor(FGSEA_subtype_list$pathway,unique(FGSEA_subtype_list$pathway))
FGSEA_subtype_list$Database=factor(FGSEA_subtype_list$Database,levels=c("Hallmark_dataset","C2_dataset"))

FGSEA_subtype_list$Group="Subtype2 vs subtype1"

Plot=Enrich_dotplot_bydataset(FGSEA_subtype_list)

write.csv(as.data.frame(FGSEA_subtype_list[,c(1:6)]),file=
             file.path(Out_table_path,"CHKIRC_NES_subtype2_Vs_type1.csv" ),row.names = FALSE)
```


```{r,fig.height=16,fig.width=20,include=T,out.width="80%"}
Plot
```



```{r}
pathwaylist1= fgsea::gmtPathways(GMT_files[1])
pathwaylist2= fgsea::gmtPathways(GMT_files[2])

pathwaylist_all=c(pathwaylist1,pathwaylist2)
stats <- wilcox_list_out_p005$log2FoldChange
names(stats) <- rownames(wilcox_list_out_p005)

Enrichment_plot_list=list()
for (name in Subtype_pathways){
  Pathway_name=name
  P=fgsea::plotEnrichment(pathwaylist_all[[Pathway_name]], stats) + labs(title=Pathway_name)
  Enrichment_plot_list[[name]]=P+theme_minimal() + theme(
    axis.text.x = element_text(size = 18, angle = 0, hjust = 1, vjust = 1),
    axis.text.y = element_text(size = 18),
    legend.position = "none"
  )
}
# 

```

```{r,fig.height=16,fig.width=28,include=T,out.width="80%"}
cowplot::plot_grid(plotlist = Enrichment_plot_list,ncol=4)
```






# Tuumor vs normal in subtype1 and 2 (wilcox test for DEGs)

```{r}
Ttest_list=list()
Subtype=unique(clinical_subtype$Subtype)%>%sort()
for ( i in Subtype){
  samples=clinical_subtype%>%dplyr::filter(Subtype==i)%>%.$Sample_ID
  Data=Pro_log_data%>%.[,samples]
  Ttest=t(Data)%>%as.data.frame() %>%apply(2,function (x) (wilcox.test(x))$p.value )%>%unlist()%>%as.data.frame()%>%setNames("pvalue") %>%mutate(padj=p.adjust(pvalue))
 Mean_log=t(Data)%>%as.data.frame()%>%apply(2,function (x) mean(x))
 Ttest_padj005=Ttest %>% mutate(log2Foldchange=Mean_log) %>%dplyr::filter(padj<0.05)
  Ttest_list[[i]]=Ttest %>% mutate(log2Foldchange=Mean_log)%>%mutate(gene_symbol=rownames(.))

}


```

```{r}
#Hallmark and C2

Fgsea_output=list()
for ( i in names(Ttest_list) ){
  res_p005=Ttest_list[[i]]%>%dplyr::filter(pvalue<0.05)
  res_p005_rank=res_p005%>%arrange(log2Foldchange) %>%.[c("log2Foldchange")]%>%as.data.frame() %>%tibble::rownames_to_column()%>%na.omit()
  
  ranks <- tibble::deframe(res_p005_rank)

 fgsea_list=list()
 for (k in GMT_files){
    name=gsub(".*/","",k)
    name2=gsub(".symbols.gmt.txt","",name)
    pathway= fgsea::gmtPathways(k)
    fgseaRes <- fgsea::fgseaMultilevel(pathway, ranks,
                           minSize=5,
                           maxSize=500,nPermSimple = 10000,eps = 0
                           )
    fgsea_list[[name2]]=fgseaRes%>%mutate(Group=i)

 }
 Fgsea_output[[i]]=do.call("rbind",fgsea_list)
}

FGSEA_list=do.call("rbind",Fgsea_output)
FGSEA_list$Group=factor(FGSEA_list$Group)

names(Ttest_list)
```

# NES analysis (Tumor vs nomral) for type1 and type2 (NES data in CHKIRC_subtype1_2_tumor_Vs_Normal_NES.csv)

```{r,fig.height=16,fig.width=20,include=T,out.width="80%"}
FGSEA_subtype_list=FGSEA_list%>%as.data.frame()%>%dplyr::filter(pathway %in%Subtype_pathways)%>%arrange(NES)%>%mutate(Database=ifelse(grepl("HALLMARK",pathway),"Hallmark_dataset","C2_dataset"))%>%mutate(State=ifelse(pathway%in%c("WP_GLYCOLYSIS_AND_GLUCONEOGENESIS"        , "WP_MITOCHONDRIAL_LONG_CHAIN_FATTY_ACID_BETAOXIDATION"),"low pathways","high pathways"))%>%arrange(pathway)
FGSEA_subtype_list$pathway=factor(FGSEA_subtype_list$pathway,unique(FGSEA_subtype_list$pathway))
FGSEA_subtype_list$Database=factor(FGSEA_subtype_list$Database,levels=c("Hallmark_dataset","C2_dataset"))

Output_list=list()
for (i in unique(FGSEA_subtype_list$State)){
  FGSEA_subtype_get=FGSEA_subtype_list%>%dplyr::filter(State==i)
 Output_list[[i]] = Enrich_dotplot_bydataset(FGSEA_subtype_get) 

  
}

 
cowplot::plot_grid(plotlist = Output_list,ncol=1)

write.csv(FGSEA_subtype_list[,c(1:6,9)],file.path(Out_table_path,"CHKIRC_subtype1_2_tumor_Vs_Normal_NES.csv"),row.names = F)
```

## Tumor vs Normal in Subtype1

```{r}
Group_set="CHKIRC_subtype1"
wilcox_list_out_p005=Ttest_list[[Group_set]]%>%dplyr::filter(pvalue<0.05)%>%dplyr::filter(!is.na(log2Foldchange))
stats <- wilcox_list_out_p005$log2Foldchange
names(stats) <- rownames(wilcox_list_out_p005)
Enrichment_plot_list=list()

for (name in Subtype_pathways){
  Pathway_name=name
  FGSEA_out=FGSEA_subtype_list%>%dplyr::filter(pathway==Pathway_name)%>%dplyr::filter(Group==Group_set)
  NES_value=FGSEA_out$NES
  padj_value=FGSEA_out$padj
# add \n to thte title
  Title <- paste0(Pathway_name, "\nNES=", round(NES_value, 2), ", padj=", round(padj_value, 3))
  
  P=fgsea::plotEnrichment(pathwaylist_all[[Pathway_name]], stats) + labs(title=Title)
  
  Enrichment_plot_list[[name]]=P+theme_minimal() + theme(
    axis.text.x = element_text(size = 18, angle = 0, hjust = 1, vjust = 1),
    axis.text.y = element_text(size = 18),
    legend.position = "none"
  )
}

```


```{r,fig.height=16,fig.width=28,include=T,out.width="80%"}
cowplot::plot_grid(plotlist = Enrichment_plot_list,ncol=4) 
```

## Tumor vs Normal in Subtype2

```{r}
Group_set="CHKIRC_subtype2"
wilcox_list_out_p005=Ttest_list[[Group_set]]%>%dplyr::filter(pvalue<0.05)%>%dplyr::filter(!is.na(log2Foldchange))
stats <- wilcox_list_out_p005$log2Foldchange
names(stats) <- rownames(wilcox_list_out_p005)
Enrichment_plot_list=list()

for (name in Subtype_pathways){
  Pathway_name=name
  FGSEA_out=FGSEA_subtype_list%>%dplyr::filter(pathway==Pathway_name)%>%dplyr::filter(Group==Group_set)
  NES_value=FGSEA_out$NES
  padj_value=FGSEA_out$padj
# add \n to thte title
  Title <- paste0(Pathway_name, "\nNES=", round(NES_value, 2), ", padj=", round(padj_value, 3))
  
  P=fgsea::plotEnrichment(pathwaylist_all[[Pathway_name]], stats) + labs(title=Title)
  
  Enrichment_plot_list[[name]]=P+theme_minimal() + theme(
    axis.text.x = element_text(size = 18, angle = 0, hjust = 1, vjust = 1),
    axis.text.y = element_text(size = 18),
    legend.position = "none"
  )
}

```


```{r,fig.height=16,fig.width=28,include=T,out.width="80%"}
cowplot::plot_grid(plotlist = Enrichment_plot_list,ncol=4)
```


```{r}
# save DEGs in tumor vs normal in subtype1 and 2
# filter ttest by pvalue<0.05 and log2FoldChange>0.27
Subtype1_DEGs=Ttest_list[[1]]%>%dplyr::filter(pvalue<0.05 )%>%dplyr::filter(log2Foldchange>0.27|log2Foldchange<as.numeric(-0.27))%>%.[,c("gene_symbol","log2Foldchange","pvalue")]

Subtype2_DEGs=Ttest_list[[2]]%>%dplyr::filter(pvalue<0.05 )%>%dplyr::filter(log2Foldchange>0.27|log2Foldchange<as.numeric(-0.27))%>%.[,c("gene_symbol","log2Foldchange","pvalue")]
# using excel to save Ttest_out_sig
library(openxlsx)

write.xlsx(Subtype1_DEGs,file.path(Out_table_path,"CHKIRC_subtype1_DEGs.xlsx"),sheetName = "Subtype1_tumor_vs_noraml_DEGs",row.names = F)
write.xlsx(Subtype2_DEGs,file.path(Out_table_path,"CHKIRC_subtype2_DEGs.xlsx"),sheetName = "Subtype2_tumor_vs_normal_DEGs",row.names = F)

library(openxlsx)

# Create workbook
wb <- createWorkbook()

# Subtype1 Sheet
addWorksheet(wb, sheetName = "Subtype1_tumor_vs_normal_DEGs")
writeData(wb, sheet = "Subtype1_tumor_vs_normal_DEGs",
          x = "Significant DEGs for Subtype1 (Tumor vs Normal, pvalue < 0.05, |log2FoldChange| > 0.27)",
          startRow = 1, colNames = FALSE)
writeData(wb, sheet = "Subtype1_tumor_vs_normal_DEGs",
          x = Subtype1_DEGs, startRow = 2, rowNames = FALSE)

# Subtype2 Sheet
addWorksheet(wb, sheetName = "Subtype2_tumor_vs_normal_DEGs")
writeData(wb, sheet = "Subtype2_tumor_vs_normal_DEGs",
          x = "Significant DEGs for Subtype2 (Tumor vs Normal, pvalue < 0.05, |log2FoldChange| > 0.27)",
          startRow = 1, colNames = FALSE)
writeData(wb, sheet = "Subtype2_tumor_vs_normal_DEGs",
          x = Subtype2_DEGs, startRow = 2, rowNames = FALSE)

# Save the workbook
saveWorkbook(wb, file = file.path(Out_table_path, "CHKIRC_all_subtypes_DEGs.xlsx"), overwrite = TRUE)

```

```{r}
knitr::knit_exit()
```
